Moved timeline to particular file (#267)
* Moved timeline to Timeline.js

* Deleted useless code and added import in Timeline.js

* Added comments

* Added comment

* Added comment

* Deleted comments

* Add sample code

* Fixed alpha channel of current reservations

* Deleted code for timeline in room list

* Partial PUT (PATCH) (#265)

* add PATCH for /users/{id} and /rooms/{id}

* Add docker-compose from sources, local database (#233)

* Add docker-compose from sources

* Change WORKDIR

* Use absolute WORKDIR

* Add database

* Add default values

* add prune in npm

* remove depends for server(backend) on database(db)

* bring back depends-on db for server

---------


* Replace alerts to snackbars (#268)

* Added context and one usage of snackbar

* Fixed snackbar context

* Move SnackbarProvider to App

* Move Snackbar from Room.js to SnackbarAlert.js

* Replaced alerts to snackbars everywhere

* deleted useless snackbar

---------


* Replace alerts to snackbars (#268)

* Added context and one usage of snackbar

* Fixed snackbar context

* Move SnackbarProvider to App

* Move Snackbar from Room.js to SnackbarAlert.js

* Replaced alerts to snackbars everywhere

* deleted useless snackbar

---------


---------
